+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -254,6 +254,8 @@ where
         answer_callback_query,
         set_my_commands,
         get_my_commands,
+        set_my_name,
+        get_my_name,
         set_my_description,
         get_my_description,
         set_my_short_description,
@@ -708,6 +710,10 @@ trait ErasableRequester<'a> {
     ) -> ErasedRequest<'a, SetMyCommands, Self::Err>;
 
     fn get_my_commands(&self) -> ErasedRequest<'a, GetMyCommands, Self::Err>;
+
+    fn set_my_name(&self) -> ErasedRequest<'a, SetMyName, Self::Err>;
+
+    fn get_my_name(&self) -> ErasedRequest<'a, GetMyName, Self::Err>;
 
     fn set_my_description(&self) -> ErasedRequest<'a, SetMyDescription, Self::Err>;
 
@@ -1510,6 +1516,14 @@ where
 
     fn get_my_commands(&self) -> ErasedRequest<'a, GetMyCommands, Self::Err> {
         Requester::get_my_commands(self).erase()
+    }
+
+    fn set_my_name(&self) -> ErasedRequest<'a, SetMyName, Self::Err> {
+        Requester::set_my_name(self).erase()
+    }
+
+    fn get_my_name(&self) -> ErasedRequest<'a, GetMyName, Self::Err> {
+        Requester::get_my_name(self).erase()
     }
 
     fn set_my_description(&self) -> ErasedRequest<'a, SetMyDescription, Self::Err> {