+++ crates/teloxide-core/src/types/bot_name.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct BotName {
+    /// The bot's name
+    pub name: String,
+}